+++ spec/googleauth/user_authorizer_spec.rb
@@ -242,7 +242,7 @@ describe Google::Auth::UserAuthorizer do
     end
 
     before(:example) do
-      stub_request(:post, 'https://www.googleapis.com/oauth2/v3/token')
+      stub_request(:post, 'https://www.googleapis.com/oauth2/v4/token')
         .to_return(body: token_json, status: 200, headers: {
                      'Content-Type' => 'application/json'
                    })
@@ -270,7 +270,7 @@ describe Google::Auth::UserAuthorizer do
 
   context 'with invalid authorization code' do
     before(:example) do
-      stub_request(:post, 'https://www.googleapis.com/oauth2/v3/token')
+      stub_request(:post, 'https://www.googleapis.com/oauth2/v4/token')
         .to_return(status: 400)
     end